IOException with semantics in form of phantom type
Meaning of IOException depends on a context. In a lot of situations one
needs to create newtype wrapper for it just to add that semantical layer
on top of it. Phantom types are a way how to get around the necessity of
creating wrapper. Instead one creates phantom type, which may be empty
data declaration, and makes it instance of IOExceptionTag. Optionally
providing specific conversion to error message.